+++ tests/testthat/test_autoRecode.R
@@ -78,3 +78,41 @@ test_that("existing lookup, duplicate new cases", {
   expect_equal(lookup$oldValue, c(112, 115, 110))
   expect_equal(lookup$newValue, c(3, 6, 7))
 })
+
+test_that("Preserve NAs while applying a template", {
+  expect_df1 <- df1 <- data.frame(idteach = 11001:11010,
+                                  idclass = c(1001, NA, NA, 1002, NA, 1003:1005, NA, 1006))
+  gads1 <- import_DF(df1)
+  expect_df1$idclass <- c(1, NA, NA, 2, NA, 3:5, NA, 6)
+  expect_gads1 <- import_DF(expect_df1)
+  expect_template_intermediate <- data.frame(oldValue = 1001:1006,
+                                             newValue = 1:6)
+
+  expect_df2 <- df2 <- data.frame(idteach = 21001:21010,
+                                  idclass = c(NA, 1001, NA, 1005, NA, NA, 1007:1008, NA, 1009))
+  gads2 <- import_DF(df2)
+  template_path <- tempfile(fileext = ".csv")
+  expect_df2$idclass <- c(NA, 1, NA, 5, NA, NA, 7:8, NA, 9)
+  expect_gads2 <- import_DF(expect_df2)
+  expect_template_final <- data.frame(oldValue = 1001:1009,
+                                      newValue = 1:9)
+
+  gads1_reco <- autoRecode(GADSdat = gads1,
+                           var = "idclass",
+                           var_suffix = "",
+                           label_suffix = "",
+                           csv_path = template_path)
+  template_intermediate <- read.csv(template_path)
+  expect_equal(gads1_reco$dat, expect_gads1$dat)
+  expect_equal(template_intermediate, expect_template_intermediate)
+
+  gads2_reco <- suppressWarnings(autoRecode(GADSdat = gads2,
+                                            var = "idclass",
+                                            var_suffix = "",
+                                            label_suffix = "",
+                                            template = template_intermediate,
+                                            csv_path = template_path))
+  template_final <- read.csv(template_path)
+  expect_equal(gads2_reco$dat, expect_gads2$dat)
+  expect_equal(template_final, expect_template_final)
+})